requirements.txt
numpy
torch
tqdm
hydra-core

black
isort
pre-commit

wandb